Initialize tree cache outside of driver
Two PFS servers (and therefore, currently, two drivers) will have to be
instantiated in main.go). The driver creates a large, in-memory data
structure--the tree cache--that now needs to be shared between the two
instances to avoid memory overhead. This change allows pachd/main.go to
instantiate one tree cache and pass it to the two PFS driver instances
to share